+++ site/src/examples/dialog/Success.tsx
@@ -1,5 +1,5 @@
 import { ReactElement, useState } from "react";
-import { Button } from "@salt-ds/core";
+import { Button, useId } from "@salt-ds/core";
 import {
   Dialog,
   DialogTitle,
@@ -9,6 +9,7 @@ import {
 
 export const Success = (): ReactElement => {
   const [open, setOpen] = useState(false);
+  const id = useId();
 
   const handleRequestOpen = () => {
     setOpen(true);
@@ -24,17 +25,15 @@ export const Success = (): ReactElement => {
 
   return (
     <>
-      <Button data-testid="dialog-button" onClick={handleRequestOpen}>
-        Open success dialog
-      </Button>
+      <Button onClick={handleRequestOpen}>Open success dialog</Button>
       <Dialog
         open={open}
         onOpenChange={onOpenChange}
         status="success"
         size="small"
-        aria-labelledby="success-dialog"
+        id={id}
       >
-        <DialogTitle id="success-dialog-heading">File uploaded</DialogTitle>
+        <DialogTitle title="File uploaded" />
         <DialogContent>
           File has been successfully uploaded to the shared drive.
         </DialogContent>